+++ modules/vpn-connection/module.tf
@@ -96,12 +96,6 @@ output "customer_gateway_configuration" {
   value     = aws_vpn_connection.vpn.customer_gateway_configuration
 }
 
-# wrapped in here-document delimiters for parsing convenience
-output "customer_gateway_configuration_heredoc" {
-  sensitive = true
-  value     = "<<END_XML\n${aws_vpn_connection.vpn.customer_gateway_configuration}\nEND_XML"
-}
-
 ## Resources
 
 # Get region of default provider